fix: rename useCheckout to index
