+++ src/pages/OrgInfo.tsx
@@ -4,7 +4,7 @@ import { OrgDetails } from '../containers/OrgDetails';
 import { useMemo } from 'react';
 import { createSearchParams } from 'react-router-dom';
 import { useAccommodationsAndOffers } from 'src/hooks/useAccommodationsAndOffers.tsx';
-import { useCheckout } from 'src/hooks/useCheckout/useCheckout';
+import { useCheckout } from 'src/hooks/useCheckout';
 
 export const OrgInfo = () => {
   const { bookingInfo } = useCheckout();
